split BucketObject out into own file
